ci: fix failing license check test
The test was using a version lower than 1.22.1 (the minimum needed for the project). The lichen license checker relies on the JSON output from `go mod download`, but when the version is to low the output looks like this:

```
go: github.com/nitrictech/nitric/core@v0.0.0-20241003062412-76ea6275fb0b requires go >= 1.22.1; switching to go1.22.8
{
        "Path": "google.golang.org/grpc",
        "Version": "v1.66.0",
...
```

resulting in invalid JSON, breaking the test. Setting the go version used for the test fixes this.